incomplete unfortunately, but i'll have the complete one later, for what it's worth

+++ P1StoreApplication/P1Models/P1ModelsClass.cs
@@ -33,6 +33,6 @@ public class P1OH_ModelsClass//Instantiate these values, they will be redefined
     public int OrderID {get; set;} = 1;
     public int fk_StoreID {get; set;} = 1;
     public int fk_CustomerID {get; set;} = 1;
-    public int total_cost {get; set;} = 0;
+    public decimal total_cost {get; set;} = 0;
     
 }
